Update Find-FIDOKey.ps1 Removed default view, added Pipeline Inuput, and more
If you don't use the -View parameter it will default it to a regular output so that you can pipe easier. Also you can now pipe in a single or an array of AAGUID and only get their information. Added -AllProperties that it will show you all the FIDO Alliance data as clean json. And added importing AAGUID through .txt, csv, and xlsx.